Add files via upload
**Problem Statement**

Design a food delivery time calculator. Each restaurant has a maximum capacity of only 7 cooking slots. Each appetiser (A) requires 1 cooking slot and every Main Course (M) requires 2 cooking slots. Appetisers require 17 minutes to be prepared and main courses take 29 minutes. Assume that the delivery time can be calculated by a simple formula that every 1km can be covered in 8 minutes.

You have to calculate how much time it would take for an order to be delivered. The orders come in and are maintained in a queue. If the restaurant cannot accommodate a new order it gets queued and the estimated time for its delivery gets calculated.

**Additional rules :**

If an order is going to take more than 2 Hr 30 minutes to get delivered, it gets denied.

- Every order has a 2 digit unique id.
- You cannot change the order of entries in the queue.
- If any order in itself cannot be prepared at once it will be denied. (Look at order 14 in the sample output)